+++ src/app/shared/footer/footer.component.spec.ts
@@ -29,22 +29,29 @@ describe('FooterComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should render a footer with a title, name and surname and year', () => {
-    const htmlFooter: HTMLElement =
-      fixture.nativeElement.querySelector('footer');
-    expect(htmlFooter).toBeTruthy();
+  it('should render a footer with a title, name and surname and year', (done) => {
+    fixture.detectChanges();
+    fixture.whenStable().then(() => {
+      const htmlFooter: HTMLElement =
+        fixture.nativeElement.querySelector('footer');
+      expect(htmlFooter).toBeTruthy();
 
-    expect(component.currentYear).toBe(new Date().getFullYear());
-    expect(component.nameAndSurname()).toBe('mock-name mock-surname');
+      console.log('footer', htmlFooter);
 
-    const htmlFooterTitle: HTMLElement =
-      fixture.nativeElement.querySelector('.footer__title');
-    expect(htmlFooterTitle.innerHTML).toBe('footer.title');
+      expect(component.currentYear).toBe(new Date().getFullYear());
+      expect(component.nameAndSurname()).toBe('mock-name mock-surname');
 
-    const htmlCopyrightHtml: HTMLElement =
-      fixture.nativeElement.querySelector('.footer__copyright');
-    expect(htmlCopyrightHtml.innerHTML).toBe(
-      `© ${component.currentYear} - ${component.nameAndSurname()}`
-    );
+      const htmlFooterTitle: HTMLElement = fixture.nativeElement.querySelector(
+        '.footer-content__title'
+      );
+      expect(htmlFooterTitle.innerHTML).toBe('footer.title');
+
+      const htmlCopyrightHtml: HTMLElement =
+        fixture.nativeElement.querySelector('.footer__copyright');
+      expect(htmlCopyrightHtml.innerHTML).toBe(
+        `© ${component.currentYear} - ${component.nameAndSurname()}`
+      );
+      done();
+    });
   });
 });